Move directory related method to Path parent class
These methods shared within Path subclasses.
