+++ src/main/java/seedu/address/model/path/Path.java
@@ -8,10 +8,14 @@ import java.util.stream.Collectors;
 
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.util.JsonUtil;
+import seedu.address.model.id.GroupId;
+import seedu.address.model.id.StudentId;
+import seedu.address.model.id.exceptions.InvalidIdException;
 import seedu.address.model.path.element.PathElement;
 import seedu.address.model.path.element.PathElementType;
 import seedu.address.model.path.element.exceptions.InvalidPathElementException;
 import seedu.address.model.path.exceptions.InvalidPathException;
+import seedu.address.model.path.exceptions.UnsupportedPathOperationException;
 /**
  * Represents a path in our application.
  */
@@ -97,6 +101,72 @@ public class Path {
             destination.add(element);
         }
         logger.info(destination.toString());
+    }
+
+    /**
+     * Checks whether the path represents a group directory.
+     *
+     * @return {@code true} if the path is a group directory; {@code false} otherwise.
+     */
+    public boolean isGroupDirectory() {
+        PathElement lastElement = this.pathElements.get(this.pathElements.size() - 1);
+        return lastElement.getType() == PathElementType.GROUPID;
+    }
+
+    /**
+     * Checks whether the path represents a student directory.
+     *
+     * @return {@code true} if the path is a student directory; {@code false} otherwise.
+     */
+    public boolean isStudentDirectory() {
+        PathElement lastElement = this.pathElements.get(this.pathElements.size() - 1);
+        return lastElement.getType() == PathElementType.STUDENTID;
+    }
+
+    /**
+     * Checks whether the path represents a root directory.
+     *
+     * @return {@code true} if the path is a root directory; {@code false} otherwise.
+     */
+    public boolean isRootDirectory() {
+        PathElement lastElement = this.pathElements.get(this.pathElements.size() - 1);
+        return lastElement.getType() == PathElementType.ROOT;
+    }
+
+    /**
+     * Retrieves the student ID from the path directory.
+     *
+     * @return The student ID.
+     * @throws UnsupportedPathOperationException If the operation is not supported based on the directory's state.
+     * @throws InvalidIdException If the retrieved ID is invalid.
+     */
+    public StudentId getStudentId() throws UnsupportedPathOperationException, InvalidIdException {
+        if (this.isGroupDirectory()) {
+            throw new UnsupportedPathOperationException("Can't find student ID in the group directory");
+        } else if (this.isRootDirectory()) {
+            throw new UnsupportedPathOperationException("Can't find student ID in the root directory");
+        }
+
+        StudentId id = StudentId.createStudentId(this.pathElements.get(2).toString());
+
+        return id;
+    }
+
+    /**
+     * Retrieves the group ID from the path directory.
+     *
+     * @return The group ID.
+     * @throws UnsupportedPathOperationException If the operation is not supported based on the directory's state.
+     * @throws InvalidIdException If the retrieved ID is invalid.
+     */
+    public GroupId getGroupId() throws UnsupportedPathOperationException, InvalidIdException {
+        if (this.isRootDirectory()) {
+            throw new UnsupportedPathOperationException("Can't find group ID in the root directory");
+        }
+
+        GroupId id = GroupId.createGroupId(this.pathElements.get(1).toString());
+
+        return id;
     }
 
     @Override
